+++ rust/crd/src/lib.rs
@@ -129,8 +129,6 @@ impl FlaskAppConfigOptions for SupersetConfigOptions {
     }
 }
 
-pub const HTTP_PORT: &str = "http";
-
 #[derive(Clone, CustomResource, Debug, Deserialize, JsonSchema, PartialEq, Serialize)]
 #[kube(
     group = "superset.stackable.tech",