+++ CHANGELOG.md
@@ -75,8 +75,10 @@
 
 - `operator-rs` `0.27.1` -> `0.31.0` ([#306], [#297], [#311])
 - Fixed the RoleGroup `selector`. It was not used before. ([#306])
-- Updated stackable image versions ([#295]).
-- [BREAKING] Use Product image selection instead of version. `spec.version` has been replaced by `spec.image` ([#304]).
+- Updated stackable image versions ([#295])
+- [BREAKING] Use Product image selection instead of version ([#304])
+  - `spec.version` has been replaced by `spec.image`
+  - `spec.statsdExporterVersion` has been removed, the statsd-exporter is now part of the images itself
 - Refactored LDAP authentication handling to use functionality from the `LdapAuthenticationProvider` ([#311])
 
 [#306]: https://github.com/stackabletech/superset-operator/pull/306