+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 ### Changed
 
 - `vector` `0.26.0` -> `0.31.0` ([#391]).
-- `operator-rs` `0.44.0` -> `0.51.1` ([#390], [#407]).
+- `operator-rs` `0.44.0` -> `0.52.0` ([#390], [#407]).
 - BREAKING: Removed SupersetDB object, since it created some problems when reinstalling or upgrading a Superset cluster. Instead, the initialization of the database was moved to the startup phase of each Superset pod. To make sure the initialization does not run in parallel, the `PodManagementPolicy` was set to `OrderedReady` and liveness/readiness probes were added. The `.spec.clusterConfig.loadExamplesOnInit` option was removed from the CRD, because loading the examples at every startup caused problems in certain scenarios, e.g. after an upgrade from Superset 1.5.3 to 2.1.0 ([#396]).
 
 ### Fixed