+++ CHANGELOG.md
@@ -6,22 +6,29 @@
 
 - Default stackableVersion to operator version ([#390]).
 - Support PodDisruptionBudgets ([#407]).
+- Added support for versions 2.1.1, 3.0.1 ([#415]).
 
 ### Changed
 
-- `vector` `0.26.0` -> `0.31.0` ([#391]).
-- `operator-rs` `0.44.0` -> `0.52.1` ([#390], [#407]).
+- `vector` `0.26.0` -> `0.33.0` ([#391], [#415]).
+- `operator-rs` `0.44.0` -> `0.55.0` ([#390], [#407], [#415]).
 - BREAKING: Removed SupersetDB object, since it created some problems when reinstalling or upgrading a Superset cluster. Instead, the initialization of the database was moved to the startup phase of each Superset pod. To make sure the initialization does not run in parallel, the `PodManagementPolicy` was set to `OrderedReady` and liveness/readiness probes were added. The `.spec.clusterConfig.loadExamplesOnInit` option was removed from the CRD, because loading the examples at every startup caused problems in certain scenarios, e.g. after an upgrade from Superset 1.5.3 to 2.1.0 ([#396]).
 
 ### Fixed
 
 - BREAKING: Rename Service port name from `superset` to `http` for consistency reasons. This change should normally not be breaking, as we only change the name, not the port. However, there might be some e.g. Ingresses that rely on the port name and need to be updated ([#394]).
+- Fixed config override support ([#415]).
+
+### Removed
+
+- Removed support for versions 1.3.2, 1.4.1, 1.4.2, 1.5.1, 1.5.3, 2.0.1 ([#415]).
 
 [#390]: https://github.com/stackabletech/superset-operator/pull/390
 [#391]: https://github.com/stackabletech/superset-operator/pull/391
 [#394]: https://github.com/stackabletech/superset-operator/pull/394
 [#396]: https://github.com/stackabletech/superset-operator/pull/396
 [#407]: https://github.com/stackabletech/superset-operator/pull/407
+[#415]: https://github.com/stackabletech/superset-operator/pull/415
 
 ## [23.7.0] - 2023-07-14