+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [23.11.0] - 2023-11-24
+
 ### Added
 
 - Default stackableVersion to operator version ([#390]).